+++ plugin/src/main/kotlin/org/jlleitschuh/gradle/ktlint/KtlintExtension.kt
@@ -84,6 +84,15 @@ internal constructor(
      */
     val additionalEditorconfigFile: RegularFileProperty = newFileProperty(objectFactory, projectLayout)
 
+    /**
+     * Disable particular rules, by default enabled in ktlint, using rule id.
+     *
+     * @since ktlint `0.34.2`
+     */
+    val disabledRules: SetProperty<String> = objectFactory.setProperty {
+        set(emptySet())
+    }
+
     private val kscriptExtension = KScriptExtension(kotlinScriptAdditionalPathApplier)
 
     /**